refactor: Dropbox dependencies
