+++ src/App.js
@@ -1,21 +1,17 @@
-/* global process */
-
 import React, { PureComponent } from 'react';
 
 import { Provider } from 'react-redux';
 import Store from './store';
-import { Dropbox } from 'dropbox';
+import parseQueryString from './util/parse_query_string';
 import {
   readInitialState,
   loadSettingsFromConfigFile,
   subscribeToChanges,
-  persistField,
   getPersistedField,
 } from './util/settings_persister';
 
 import runAllMigrations from './migrations';
-import parseQueryString from './util/parse_query_string';
-import { BrowserRouter } from 'react-router-dom';
+import { BrowserRouter as Router } from 'react-router-dom';
 
 import { DragDropContext } from 'react-beautiful-dnd';
 
@@ -71,27 +67,6 @@ const handleGitLabAuthResponse = async (oauthClient) => {
   }
 };
 
-/* BEGIN: Dropbox helper functions */
-
-function getCodeFromUrl() {
-  return parseQueryString(window.location.search).code;
-}
-
-function createDropboxClient(dbxAuth, client, initialState) {
-  var tmpClient = new Dropbox({
-    auth: dbxAuth,
-  });
-
-  client = createDropboxSyncBackendClient(tmpClient);
-  initialState.syncBackend = Map({
-    isAuthenticated: true,
-    client,
-  });
-  return client;
-}
-
-/* END: Dropbox helper functions */
-
 export default class App extends PureComponent {
   constructor(props) {
     super(props);
@@ -106,37 +81,11 @@ export default class App extends PureComponent {
     if (!!authenticatedSyncService) {
       switch (authenticatedSyncService) {
         case 'Dropbox':
-          const REDIRECT_URI = window.location.origin + '/';
-
-          const dbx = new Dropbox({
-            clientId: process.env.REACT_APP_DROPBOX_CLIENT_ID,
-            fetch: fetch.bind(window),
+          client = createDropboxSyncBackendClient();
+          initialState.syncBackend = Map({
+            isAuthenticated: true,
+            client: client,
           });
-          const dbxAuth = dbx.auth;
-
-          if (getCodeFromUrl()) {
-            dbxAuth.setCodeVerifier(getPersistedField('codeVerifier'));
-            dbxAuth
-              .getAccessTokenFromCode(REDIRECT_URI, getCodeFromUrl())
-              .then((response) => {
-                const dropboxAccessToken = response.result.access_token;
-
-                dbxAuth.setAccessToken(dropboxAccessToken);
-                persistField('dropboxAccessToken', dropboxAccessToken);
-
-                client = createDropboxClient(dbxAuth, client, initialState);
-
-                window.location.href = '/files';
-              })
-              .catch((error) => {
-                console.error(error);
-              });
-          } else {
-            const dropboxAccessToken = getPersistedField('dropboxAccessToken');
-            dbxAuth.setCodeVerifier(getPersistedField('codeVerifier'));
-            dbxAuth.setAccessToken(dropboxAccessToken);
-            client = createDropboxClient(dbxAuth, client, initialState);
-          }
           break;
         case 'GitLab':
           const gitlabOAuth = createGitlabOAuth();
